NGWPC PI-10 Delivery: 7716 7717 10980 10984 (NHF Modifications)

+++ src/hydrofabric_builds/hydrofabric/graph.py
@@ -166,9 +166,9 @@ def _build_upstream_dict_from_nexus(
     """
     fp_pl = flowpaths_pl.with_columns(
         [
-            pl.col(edge_id).cast(pl.Int32),
-            pl.col(f"up_{node_id}").cast(pl.Int32),
-            pl.col(f"dn_{node_id}").cast(pl.Int32),
+            pl.col(edge_id).cast(pl.Int64),
+            pl.col(f"up_{node_id}").cast(pl.Int64),
+            pl.col(f"dn_{node_id}").cast(pl.Int64),
         ]
     )
     # Create mapping: nex_id -> downstream fp_id (where this nexus is the upstream nexus)
